requirements.txt
apache-airflow==2.2.3
pendulum==2.1.2
pandas
requests
sqlalchemy
psycopg2-binary==2.9.3
python-dotenv
redis==3.5.3

